Swap - Rework the getProviders logic for better privacy
