+++ src/renderer/reducers/application.js
@@ -3,16 +3,7 @@
 import { handleActions } from "redux-actions";
 import { getSystemLocale } from "~/helpers/systemLocale";
 import { getLanguages } from "~/config/languages";
-import { createSelector } from "reselect";
-import type { OutputSelector } from "reselect";
 import type { LangAndRegion } from "~/renderer/reducers/settings";
-import type { AvailableProvider } from "@ledgerhq/live-common/lib/swap/types";
-import type { TokenCurrency, CryptoCurrency } from "@ledgerhq/live-common/lib/types";
-import { isCurrencySwapSupported } from "@ledgerhq/live-common/lib/swap";
-import type { State } from ".";
-import uniq from "lodash/uniq";
-import { isCurrencySupported } from "@ledgerhq/live-common/lib/currencies";
-import { findCryptoCurrencyById, findTokenById } from "@ledgerhq/cryptoassets";
 
 export type ApplicationState = {
   isLocked?: boolean,
@@ -21,7 +12,6 @@ export type ApplicationState = {
   osDarkMode?: boolean,
   osLanguage?: LangAndRegion,
   navigationLocked?: boolean,
-  swapProviders?: AvailableProvider[],
 };
 
 const { language, region } = getSystemLocale();
@@ -35,7 +25,6 @@ const state: ApplicationState = {
     region: osLangSupported ? region : "US",
     useSystem: true,
   },
-  swapProviders: undefined,
   hasPassword: false,
   dismissedCarousel: false,
 };
@@ -63,31 +52,6 @@ export const osDarkModeSelector = (state: Object) => state.application.osDarkMod
 export const osLangAndRegionSelector = (state: Object) => state.application.osLanguage;
 
 export const isNavigationLocked = (state: Object) => state.application.navigationLocked;
-
-export const swapProvidersSelector = (state: Object) => state.application.swapProviders;
-
-export const swapSupportedCurrenciesSelector: OutputSelector<
-  State,
-  { accountId: string },
-  (TokenCurrency | CryptoCurrency)[],
-> = createSelector(swapProvidersSelector, swapProviders => {
-  if (!swapProviders) return [];
-
-  const allIds = uniq(
-    swapProviders.reduce((ac, { supportedCurrencies }) => [...ac, ...supportedCurrencies], []),
-  );
-
-  const tokenCurrencies = allIds
-    .map(findTokenById)
-    .filter(Boolean)
-    .filter(t => !t.delisted);
-  const cryptoCurrencies = allIds
-    .map(findCryptoCurrencyById)
-    .filter(Boolean)
-    .filter(isCurrencySupported);
-
-  return [...cryptoCurrencies, ...tokenCurrencies].filter(isCurrencySwapSupported);
-});
 
 // Exporting reducer
