+++ src/renderer/screens/account/AccountHeaderActions.js
@@ -6,7 +6,7 @@ import { useSelector, connect } from "react-redux";
 import { withTranslation, Trans } from "react-i18next";
 import styled from "styled-components";
 import type { Account, AccountLike } from "@ledgerhq/live-common/lib/types";
-import { swapSupportedCurrenciesSelector } from "~/renderer/reducers/application";
+import { swapSupportedCurrenciesSelector } from "~/renderer/reducers/settings";
 import Tooltip from "~/renderer/components/Tooltip";
 import {
   isAccountEmpty,
